+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.3] - 2023-11-21
+
+### Changed
+
+- Display playist images
+- Link to playlists
+- Some more style tweaks
+- Add Spotify attribution
+
 ## [0.5.2] - 2023-11-21
 
 ### Changed